backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9

# Database (meta store)
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# DB drivers (managed connections)
psycopg2-binary==2.9.9
PyMySQL==1.1.0
aiomysql==0.2.0
clickhouse-driver==0.2.7
clickhouse-sqlalchemy==0.3.1

# Task queue
celery==5.4.0
redis==5.0.4
flower==2.0.1

# Utilities
httpx==0.27.0
python-dotenv==1.0.1
psutil==5.9.8
aiofiles==23.2.1
